+++ changelog.md
@@ -3,6 +3,7 @@
 ### 1.5
 - Fix for region with non-US date formats
 - Fix for scrolling on Events and Camps page
+- Fix for Android navigation bar
 
 ### 1.3
 - Map of Restrooms added